output/csv: fix dedup logic
This fixes the dedup logic in the csv output to:
- dedup across samples (much like the vcd output module)
- not print two values for every group of samples

`previous_sample` was being freed every receive, and so it was
impossible to dedup across multiple receives, which would cause many
rows even when the level hasn't changed when using high sample rates.

There was also a range check for `i` that makes it so that two values
will always skip the dedup check, again spamming rows even when the
level hasn't changed.