rigol-ds: Add support for DSO3000 USB comms
This is the protocol used by Agilent DSO3000 series (likely the same as
Rigol DS5000 -- the device has the same USB ID).

The USB comms on the scope is rather flaky (as is about anything in the
firmware) and undocumented and the vendor suggests it actually shouldn't
be used by anything other than their tool where they probably just
worked around the issues. This aims to do the same and seem to do the
trick in practice.