francaise-instrumentation-ams515: Cleanup, and try to resync serial
If the device is powered cycled, it reverts to echo mode which messes up
communication, so we try to detect this and disable echo again.

Channel B is actually 3A max.

We now also implement overcurrent detection, however in this state it
refuses to answer other commands so it just becomes unusable.

And since it doesn't send anything by itself we will need to use a timer
to poll for changes someday.