siglent-sds: Add SDS2000X Plus series support
Add support for both analog and digital (logic analyzer) channels.

Don't set num_block_bytes before any waveform data is read.

Add a pause after sending a command to enter ARM mode. This is necessary to prevent crashes when the timebase is large.

Move the start point after every full block is read.

Don't wait for the instrument to prepare its output buffers if it's a SDS2000X Plus.

Fix multi-channel support and ensure that the correct sample rate is obtained when the timebase is changed or when channels are enabled/disabled in "display" mode.

If a channel is turned on/off on the oscilloscope then use sigrok's settings instead.