configure: Accept glibmm-2.68 as an alternative to glibmm-2.4
Check for both glibmm-2.4 as well as glibmm-2.68. Prefer the former
when either is available, to remain backwards compatible.

This unbreaks builds on OS X with homebrew where only glibmm-2.68
is available.

Implementor's note: It's yet to get determined whether glibmm-2.68
is to get preferred, being the first stable release of that external
dependency. https://gitlab.gnome.org/GNOME/glibmm/-/blob/2.68.2/NEWS
The commit is phrased to simplify future adjustment of this preference.

This commit is based on work that was
Submitted-By: Fabrice Fontaine <fontaine.fabrice@gmail.com>