dmm/metex14: fix values for digits properties
Unify the metex14 package parser with the other drivers regarding the use of
the digits properties in the analog payload. This commit adjusts the previously
unclear, but now clarified, use of these properties:

Number of significant digits after the decimal point, if positive. When
negative, exponent with reversed polarity that is necessary to express the
value with all digits without a decimal point.

This commit partly reverts commit 7fb4ff0